Write out the github event as a file in working dir
The action being performed for a web hook may want to do more complex
checking of the event information to decide on an action. We don't want
to put all of this in the environment so we write out the full event to
a file. The handlers can then parse this file if they wish.